Modify deployment code
