+++ crosschain-resolver/deploy_l1/10_l1resolver.ts
@@ -9,12 +9,11 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const {deployer} = await getNamedAccounts();
 
   const OP_VERIFIER_ADDRESS = process.env.OP_VERIFIER_ADDRESS
-  const OwnedResolver = await hre.companionNetworks['l2'].deployments.get('OwnedResolver');
   if(!OP_VERIFIER_ADDRESS) throw ('Set $OP_VERIFIER_ADDRESS')
-  console.log({OP_VERIFIER_ADDRESS, OWNED_RESOLVER_ADDRESS:OwnedResolver.address})
+  console.log({OP_VERIFIER_ADDRESS})
   await deploy('L1Resolver', {
     from: deployer,
-    args: [OP_VERIFIER_ADDRESS, OwnedResolver.address],
+    args: [OP_VERIFIER_ADDRESS],
     log: true,
   });
 };
